# Conflicts:
#	finn-rtllib/softmax_rtl/binopf.sv
#	finn-rtllib/softmax_rtl/int_to_fp32.sv
#	finn-rtllib/softmax_rtl/softmaxf.sv
#	finn-rtllib/softmax_rtl/softmaxf_pkg.sv
#	src/finn/builder/build_dataflow.py
#	src/finn/builder/build_dataflow_config.py
#	src/finn/builder/build_dataflow_steps.py
#	src/finn/custom_op/fpgadataflow/rtl/__init__.py
#	src/finn/custom_op/fpgadataflow/rtl/elementwise_binary_rtl.py
#	src/finn/custom_op/fpgadataflow/rtl/hwsoftmax_rtl.py
#	src/finn/transformation/fpgadataflow/specialize_layers.py
#	src/finn/util/fpgadataflow.py
#	tests/fpgadataflow/test_fpgadataflow_softmax.py

+++ src/finn/transformation/qonnx/qonnx_activation_handlers.py
@@ -367,11 +367,15 @@ class QuantReluHandler(QuantActBaseHandler):
             # onnx/finn/handler/act.py#L21
             num_distinct_values = 2**bit_width
             num_thresholds = int(num_distinct_values - 1)
-            flat_scale = quant_scale.flatten().astype(np.float32)
+            # Accumulate in float64 and narrow to the storage dtype once at the
+            # end, so the per-threshold value carries at most a single float32
+            # rounding step rather than accumulated error across min_threshold +
+            # step*t (see QuantIdentityHandler._calculate_thresholds).
+            flat_scale = quant_scale.flatten().astype(np.float64)
             num_scale_channels = flat_scale.shape[0]
-            step = np.abs(flat_scale).astype(np.float32)
+            step = np.abs(flat_scale)
             min_threshold = step / 2
-            thresholds = np.empty((num_scale_channels, num_thresholds), dtype=np_default_dtype)
+            thresholds = np.empty((num_scale_channels, num_thresholds), dtype=np.float64)
             for c in range(num_scale_channels):
                 for t in range(num_thresholds):
                     thresholds[c][t] = min_threshold[c] + step[c] * t
@@ -385,15 +389,15 @@ class QuantReluHandler(QuantActBaseHandler):
                 num_distinct_values = 2**bit_width
 
             num_thresholds = int(num_distinct_values - 1)
-            flat_scale = quant_scale.flatten().astype(np.float32)
+            flat_scale = quant_scale.flatten().astype(np.float64)
             num_scale_channels = flat_scale.shape[0]
-            scale = np.abs(flat_scale).astype(np.float32)
+            scale = np.abs(flat_scale)
             half_scale = scale / 2
             # alpha and lambda
             # from https://pytorch.org/docs/stable/generated/torch.nn.SELU.html
             alpha = 1.6732632423543772848170429916717
             selu_scale = 1.0507009873554804934193349852946
-            thresholds = np.empty((num_scale_channels, num_thresholds), dtype=np_default_dtype)
+            thresholds = np.empty((num_scale_channels, num_thresholds), dtype=np.float64)
             for c in range(num_scale_channels):
                 for t in range(num_thresholds):
                     step = -1.0 + half_scale + scale[c] * t
@@ -401,6 +405,9 @@ class QuantReluHandler(QuantActBaseHandler):
                         thresholds[c][t] = np.log(step / (alpha * selu_scale) + 1)
                     else:
                         thresholds[c][t] = step / selu_scale
+
+        # narrow back to the storage dtype only after the accumulation
+        thresholds = thresholds.astype(np_default_dtype)
 
         # First try to consider the tensor layout of the output for determining
         # the number of output channels
@@ -552,11 +559,17 @@ class QuantIdentityHandler(QuantActBaseHandler):
                 num_distinct_values = 2**bit_width
 
             num_thresholds = int(num_distinct_values - 1)
-            flat_scale = quant_scale.flatten()
+            # Accumulate the thresholds in float64. Each threshold is computed as
+            # a large base (min_threshold ~ -step*num_thresholds/2) plus a large
+            # multiple (step*t); doing this in float32 leaves ~1 ulp of error at
+            # the base's magnitude (~5e-7 for typical scales), which can shift a
+            # boundary below an input that Quant legitimately rounds down, causing
+            # a one-step MultiThreshold mismatch. float64 keeps that error ~1e-15.
+            flat_scale = quant_scale.flatten().astype(np.float64)
             num_scale_channels = flat_scale.shape[0]
             step = np.abs(flat_scale)
             half_step = step / 2.0
-            thresholds = np.empty((num_scale_channels, num_thresholds), dtype=np_default_dtype)
+            thresholds = np.empty((num_scale_channels, num_thresholds), dtype=np.float64)
             # compute the value of the smallest threshold, we'll neg-bias all
             # generated thresholds by this much
             min_threshold = -half_step - step * ((num_thresholds // 2) - 1)
@@ -567,6 +580,8 @@ class QuantIdentityHandler(QuantActBaseHandler):
             for c in range(num_scale_channels):
                 for t in range(num_thresholds):
                     thresholds[c][t] = min_threshold[c] + step[c] * t
+            # narrow back to the storage dtype only after the accumulation
+            thresholds = thresholds.astype(np_default_dtype)
 
             # First try to consider the tensor layout of the output for
             # determining the number of output channels
